pyproject.toml
[tool.poetry]
name = "matatika-ext"
version = "0.0.1"
description = "`matatika-ext` is a Meltano utility extension."
authors = ["Reuben Frankel"]
license = "Apache 2.0"
packages = [
    { include = "matatika_ext" },
    { include = "files_matatika_ext" },
]
exclude = [
    "**/__pycache__",
]

[tool.poetry.dependencies]
python = "<3.12,>=3.7"
PyYAML = "^6.0.0"
click = "^8.1.3"
typer = "^0.6.1"
"meltano.edk"= {git = "https://github.com/meltano/edk.git", rev="main"}
importlib-resources = "^5.12.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
matatika_extension = 'matatika_ext.main:app'
matatika_invoker = 'matatika_ext.pass_through:pass_through_cli'
